Cast to int now we have integer values, removes 4 warnings.
git-svn-id: svn://scribus.net/branches/Version13x/Scribus@5027 11d20701-8431-0410-a711-e3c959e3b870

+++ scribus/pageitem.cpp
@@ -3329,7 +3329,7 @@ bool PageItem::pointWithinItem(const int x, const int y)
 	// FIXME: We should be rounding or truncating here, not letting the compiler do it.
 	// Should we be rounding, truncating up, or truncating down?
 	//CB is this now correct?
-	QRect itemRect( floor(BoundingX), floor(BoundingY), ceil(BoundingW), ceil(BoundingH) );
+	QRect itemRect( static_cast<int>floor(BoundingX), static_cast<int>floor(BoundingY), static_cast<int>ceil(BoundingW), static_cast<int>ceil(BoundingH) );
 	return itemRect.contains(x, y);
 }
